add id to serilizer category

+++ project/django/store/market/views/cbv_like.py
@@ -8,6 +8,8 @@ from rest_framework.response import Response
 
 def getOwner(request):
     token = request.headers.get('Authorization')
+    if token is None:
+        return None
     token = token.split(' ')[1]
     owner = Token.objects.get(key=token).user
     return owner
@@ -18,30 +20,32 @@ class LikeApiView(APIView):
     def post(self, request):
         serializer = LikeSerializer(data=request.data)
         author = getOwner(request)
-        if serializer.is_valid():
-            item_id = serializer.validated_data.pop('item_id')
-            item = Item.objects.get(pk=item_id)
-            like = Like.objects.filter(item=item, author=author)
-            if like.count() == 0:
-                like = serializer.save(item_id=item_id, author=author)
-                return Response(like.as_dict())
-            return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        else:
-            return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        if author is not None:
+            if serializer.is_valid():
+                item_id = serializer.validated_data.pop('item_id')
+                item = Item.objects.get(pk=item_id)
+                like = Like.objects.filter(item=item, author=author)
+                if like.count() == 0:
+                    like = serializer.save(item_id=item_id, author=author)
+                    return Response(like.as_dict())
+                return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            else:
+                return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
 
     def delete(self, request):
         author = getOwner(request)
         serializer = LikeSerializer(data=request.data)
-        if serializer.is_valid():
-            item_id = serializer.validated_data.pop('item_id')
-            item = Item.objects.get(pk=item_id)
-            like = Like.objects.filter(item=item, author=author)
-            if like.count() != 0:
-                item.likes_count -= 1
-                item.save()
-                like = Like.objects.get(item=item, author=author)
-                like.delete()
-                return Response(like.as_dict())
-            return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
-        else:
-            return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+        if author is not None:
+            if serializer.is_valid():
+                item_id = serializer.validated_data.pop('item_id')
+                item = Item.objects.get(pk=item_id)
+                like = Like.objects.filter(item=item, author=author)
+                if like.count() != 0:
+                    item.likes_count -= 1
+                    item.save()
+                    like = Like.objects.get(item=item, author=author)
+                    like.delete()
+                    return Response(like.as_dict())
+                return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+            else:
+                return Response(serializer.errors, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
